% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthdata-package.R
\docType{package}
\name{healthdata-package}
\alias{healthdata}
\alias{healthdata-package}
\title{healthdata: Some useful health datasets}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
