% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacrc}
\alias{metacrc}
\title{Colorectal cancer survival.}
\format{
A data frame with 19 variables:
\describe{
\item{\code{id}}{patients id}
\item{\code{age}}{age at metastases}
\item{\code{sex}}{sex}
\item{\code{stage}}{stage at diagnosis}
\item{\code{APC}}{Adenomatous Polyposis Coli (APC), a gene that encodes a tumor suppressor protein}
\item{\code{type}}{Cancer type detailed}
\item{\code{location}}{Tumor location}
\item{\code{carcinomatosis}}{Carcinomatosis}
\item{\code{differentiation}}{Tumor differentiation}
\item{\code{ECOG}}{Eastern Cooperative Oncology Group score}
\item{\code{fraction}}{Fraction of genome altered}
\item{\code{LST}}{Large-scale state transition}
\item{\code{bone}}{Bone metastasis}
\item{\code{MSI}}{MSI score}
\item{\code{mutation}}{Mutation count}
\item{\code{TMB}}{TMB nonsynonymous}
\item{\code{status}}{Survival status}
\item{\code{time}}{Survival time}
}

For further details, see the original article \url{http:https://pubmed.ncbi.nlm.nih.gov/32730818/}
}
\usage{
metacrc
}
\description{
Patients diagnosed with mCRC and evaluated for palliative chemotherapy.
}
\keyword{datasets}
